% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr-class.R
\name{mtr}
\alias{mtr}
\title{mtr}
\description{
A class for multiple target regression.
}
\section{Usage}{

\preformatted{
set.seed(1)
x <- matrix(rnorm(100), ncol = 10)
set.seed(2)
y <- matrix(rnorm(30), ncol = 3)

temp <- mtr$new(regMethod = "lm")
temp$fit(x, y)
temp$predict(x)
}
}

\section{Arguments}{

\describe{
  \item{mtr}{The mtr class.}
  \item{more}{documentation required TODO.}
}
}

\section{Details}{

some details.
}

\examples{
set.seed(1)
x <- matrix(rnorm(100), ncol = 10)
set.seed(2)
y <- matrix(rnorm(30), ncol = 3)

temp <- mtr$new(regMethod = "lm")
temp$fit(x, y)
temp$predict(x)

}
