% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-target.R
\name{predict_single_target}
\alias{predict_single_target}
\title{Predict method for single target regressions.}
\usage{
predict_single_target(object, newdata)
}
\description{
Predict method for single target regressions.
}
