% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-target.R
\name{single_target}
\alias{single_target}
\title{Single target regressions.}
\usage{
single_target(x, y, regMethod, ...)
}
\description{
Decompose the multiple target regression into independent univariate regressions.
}
